% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{A3LNetwork}
\alias{A3LNetwork}
\title{The A3LNetwork graph}
\format{
\subsection{\code{A3LNetwork}}{

An igraph object containing 7 538 nodes and 53 919 directed edges.
}
}
\source{
<StrinDB: https://string-db.org/cgi/download?sessionId=bjv6RHolIynV&species_text=Homo+sapiens>
}
\usage{
A3LNetwork
}
\description{
A medium size directed biological network comprised of three layers:
\itemize{
\item proteins
\item genes
\item GO terms
The node names have prefix allowing their identication:
Prot_, Gene_, GO_
A \code{type} attribute is also present on nodes for identification based on user preference
type = protein, gene, or go
}
}
\keyword{datasets}
