% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add}
\alias{add}
\title{Concatenate with replacement}
\usage{
add(v1, x)
}
\arguments{
\item{v1}{A vector or a list}

\item{x}{The element to be added}
}
\value{
The length of the unlist between v1 and v2
}
\description{
Add an element to a vector or a list with replacement
}
\examples{
a = 1:10
add(a, 11)
}
