% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_pulldown.R
\name{calculate_relevance_score}
\alias{calculate_relevance_score}
\title{Calculate ECC Relevance Scores for Network Nodes}
\usage{
calculate_relevance_score(network, seed_nodes, database)
}
\arguments{
\item{network}{A data frame representing the interactions in the network. Expected
to have columns 'nodes' and 'interactors' denoting the nodes and their interactions, respectively.}

\item{seed_nodes}{A vector containing seed nodes which are to be used in determining relevance scores.}

\item{database}{A data frame representing the interactions in the database. Expected
to have a column named 'nodes' representing the nodes.}
}
\value{
A data frame with nodes and their respective ECC relevance scores.
}
\description{
Calculates the ECC relevance score for all nodes in the provided network based on interactions
from a database and a set of seed nodes. This function considers triangle interactions
and combines interaction counts from the network and a provided database to determine
the ECC relevance scores.
}
\examples{
\dontrun{
# Assuming you have a data frame 'my_network' representing your network,
# a vector 'my_seed_nodes' containing seed nodes, and a data frame 'my_database'
# representing the interactions database:
scores <- calculate_relevance_score(my_network, my_seed_nodes, my_database)
}
}
