% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{check_for_validity}
\alias{check_for_validity}
\title{Check arguments}
\usage{
check_for_validity(
  Graph,
  start_nodes,
  Signature_list = NULL,
  Cluster_list = NULL,
  dist_matrix = NULL
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{Signature_list}{Optional. Your list of signatures.}

\item{Cluster_list}{Optional. Your list of clusters previously calculated or not.}

\item{dist_matrix}{Optional. Your distance matrix previously calculated.}
}
\value{
Stop the called function if some errors are detected, modify some arguments if necessary.
}
\description{
Check arguments given within called functions
}
