% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS_and_ML.R
\name{check_for_validity2}
\alias{check_for_validity2}
\title{Internal function: Check for validitity of ML functions}
\usage{
check_for_validity2(
  df,
  class,
  nbr_of_boot = NULL,
  seed = NULL,
  nfolds = NULL,
  nCores = NULL,
  verbose = NULL
)
}
\arguments{
\item{df}{A numerical dataframe}

\item{class}{A boolean vector representing the class (TRUE/FALSE)}

\item{nbr_of_boot}{A natural number higher than 1, the number of bootstraps done.}

\item{seed}{seed for reproductability}

\item{nfolds}{number of folds during cross validation}

\item{nCores}{number of cores for parallel computing}

\item{verbose}{logical, wether to print or not internal calculations.}
}
\value{
Return an error, or let the function continue if everything is fine.
}
\description{
Return an error and stop the function if arguments are not how they should be
}
