% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunks}
\alias{chunks}
\title{Create chunks}
\usage{
chunks(x, n)
}
\arguments{
\item{x}{The vector}

\item{n}{The number of chunks}
}
\value{
A list of chunks from the original vector
}
\description{
Divide a vector into chunks of the same size
}
\examples{
list_of_letters = chunks(letters, 5)
}
