% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_analysis.R
\name{convert_to_entrez}
\alias{convert_to_entrez}
\title{Convert Various Gene ID Types to Entrez IDs}
\usage{
convert_to_entrez(gene_list, mart, id_type)
}
\arguments{
\item{gene_list}{A character vector representing a list of gene identifiers to be converted.}

\item{mart}{A biomaRt object that specifies the database and dataset to use for the conversion.}

\item{id_type}{A character string that indicates the type of gene ID provided in `gene_list`
(e.g., "uniprot", "ensembl", "ensembl_peptide", "ensembl_transcript", "hgnc").}
}
\value{
A character vector of converted Entrez gene IDs.
}
\description{
This function facilitates the conversion of various types of gene IDs (like "uniprot", "ensembl", etc.) to Entrez IDs.
It leverages the biomaRt package to achieve this.
}
\examples{
\dontrun{
# Assuming `sample_genes` is a predefined list of gene identifiers and `sample_mart` is a defined biomaRt object:
# entrez_ids <- convert_to_entrez(gene_list = sample_genes, mart = sample_mart, id_type = "uniprot")
}

}
