% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_by_rwr}
\alias{extract_by_rwr}
\title{Random Walk distance: to}
\usage{
extract_by_rwr(Graph, start_nodes, nCores = 1, verbose = TRUE)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{nCores}{Default: 1 The number of cores used. Keep 1 if you want sequential calculations, >1 for parallel.}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different distance features extracted from the network.
}
\description{
Calculate the random walk distance from  starting_nodes to  all other nodes in the network.
}
\examples{
# Don't run
# extract_by_rwr = extract_by_rwr(Network, c("node1", "node1"))
}
