% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_by_shp.intern}
\alias{extract_by_shp.intern}
\title{Internal function: shortest path distance}
\usage{
extract_by_shp.intern(
  Graph,
  start_nodes,
  to = vnames(Graph),
  mode = "out",
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{to}{Optional. Distance to what nodes.}

\item{mode}{Optional. The directionality of the distance measures, can be "out", "in", or "all.}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different distance features extracted from the network.
}
\description{
An internal function that calculate the shortest path distance from starting_nodes to any other nodes in the network.
}
