% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_by_shp}
\alias{extract_by_shp}
\title{Shortest path distance: to}
\usage{
extract_by_shp(Graph, start_nodes, verbose = TRUE)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different distance features extracted from the network.
}
\description{
Calculate the shortest path distance from starting_nodes to all other nodes in the network.
}
\examples{
# Don't run
# extract_by_shp = extract_by_shp(Network, c("node1", "node1"))
}
