% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_by_shp_ind}
\alias{extract_by_shp_ind}
\title{Shortest path distance}
\usage{
extract_by_shp_ind(Graph, start_nodes, verbose = TRUE)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different distance features extracted from the network.
}
\description{
Calculate the shortest path distance between starting_nodes and all other nodes in the network in a undirectional manner.
}
\examples{
# Don't run
# extract_by_shp_ind = extract_by_shp_ind(Network, c("node1", "node1"))
}
