% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_by_sig}
\alias{extract_by_sig}
\title{Extract signature-based features}
\usage{
extract_by_sig(
  Graph,
  start_nodes = vnames(Graph),
  Signature_list,
  Cluster_list = NULL,
  dist_matrix = NULL,
  return_clusters = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{Signature_list}{A named list of nodes that are in the network.}

\item{Cluster_list}{Optional. A Named list of clusters previously calculated.}

\item{dist_matrix}{Optional. A distance matrix previously calculated.}

\item{return_clusters}{Default: FALSE.  Wether or not to return the clusters utilized}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different features extracted from the network.
}
\description{
Extract signature-based features from a biological network based on a list of signatures nodes.
}
\examples{
# Don't run
# Signature = list(Signature1 = c("node7", "node9", "node13"))
# Dist_signature = extract_by_sig(Network, c("node1", "node1"), Signature)
}
