% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_cluster}
\alias{extract_cluster}
\title{Extract cluster-based features}
\usage{
extract_cluster(
  Graph,
  start_nodes = vnames(Graph),
  Cluster_list = NULL,
  dist_matrix = NULL,
  only_cluster = FALSE,
  return_clusters = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{Cluster_list}{Optional. A Named list of clusters previously calculated. If not, it will calculate its own from the network}

\item{dist_matrix}{Optional. A distance matrix previously calculated.}

\item{only_cluster}{Default: FALSE.  If TRUE, the function only calculates and return the modules identified without distance calculations.}

\item{return_clusters}{Default: FALSE.  Weather or not to return the clusters utilized}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different features extracted from the network.
}
\description{
Extract cluster-based features from a biological network, including distance to start_nodes to each clusters.
}
\examples{
# Don't run
# Dist_clusters = extract_cluster(Network, c("node1", "node1"), return_clusters = TRUE)
}
