% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_more_si_features}
\alias{extract_more_si_features}
\title{Extract signature-based features}
\usage{
extract_more_si_features(
  Graph,
  start_nodes = vnames(Graph),
  Signature_list,
  FC_vectors,
  topo_results,
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{Signature_list}{A named list of nodes that are in the network.}

\item{FC_vectors}{A named list of fold changes values for each node in the network. 0 are put where fold changes is missing}

\item{topo_results}{The results returned by the function extract_topolo}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different features extracted from the network.
}
\description{
Extract signature-based features from a biological network based on a list of signatures nodes.
}
\examples{
# Don't run
# Signature = list(Signature1 = c("node7", "node9", "node13"))
# Dist_signature = extract_more_si_features(Network, c("node1", "node1"), Signature)
}
