% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_topolo}
\alias{extract_topolo}
\title{Extract topological metrics-based features}
\usage{
extract_topolo(Graph, start_nodes = vnames(Graph), verbose = TRUE)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{verbose}{Default: TRUE. Wether or not to print the internal calculations}
}
\value{
Return a dataframe where columns are different features extracted from the network.
}
\description{
Extract topological-based features from a biological network including degree, betweenness, similarity, etc.
}
\examples{
# Don't run
# Topological_metrics = extract_topolo(Network, c("node1", "node1"))
}
