% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{get.adjacency_matrix}
\alias{get.adjacency_matrix}
\title{Internal function: calculate adjacency matrix}
\usage{
get.adjacency_matrix(graph, attr = NULL)
}
\arguments{
\item{graph}{An igraph object}

\item{attr}{A weight attributes for the edges, not functional yet.}
}
\value{
The length of the unlist between v1 and v2
}
\description{
Calculate the adjacency matrix of a network
}
