% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_pulldown.R
\name{get_all_interactions}
\alias{get_all_interactions}
\title{Retrieve All Interactions from a Network Database}
\usage{
get_all_interactions(database)
}
\arguments{
\item{database}{A data frame representing the interactions in the network.
Expected to have columns 'nodes' and 'interactors' denoting the nodes and their interactions, respectively.}
}
\value{
A data frame that includes all the interactions from the network, considering
both directions of each interaction. The returned data frame has columns 'nodes' and 'interactors'.
}
\description{
This function retrieves all the interactions from the provided network database.
Since some networks might include one-way interactions, this function ensures
that all interactions are considered to be two-way (bidirectional) by adding
reverse interactions if they are not already present.
}
\examples{
\dontrun{
# Assuming you have a data frame 'my_database' representing your network:
all_interactions <- get_all_interactions(my_database)
}
}
