% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS_and_ML.R
\name{InformationGain}
\alias{InformationGain}
\title{Calculate information gain}
\usage{
InformationGain(df, class, verbose = TRUE)
}
\arguments{
\item{df}{A numerical dataframe or matrix}

\item{class}{A boolean vector representing the class (TRUE/FALSE)}

\item{verbose}{Default = TRUE, print internal calculations.}
}
\value{
A dataframe of two columns: InfGain (information gain value) and feature (name of the feature).
}
\description{
Calculate the information gain of each features (column) in a dataframe
}
\examples{
if(requireNamespace("mlbench", quietly = TRUE)) {
data("BreastCancer", package = "mlbench")
}

# Remove rows with missing values
BreastCancer = BreastCancer[rowSums(is.na(BreastCancer)) == 0, ]

# Remove the ID column and the class column.
df = BreastCancer[, -c(1, 11)]

# Convert each column to numeric class
df = as.data.frame(apply(df, 2, as.numeric))

# Retrieve the class as a logical vector
class = BreastCancer$Class == "malignant"

# Calculate information gain for each features
BreastCancer_InfoGain = InformationGain(df, class)
}
