% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inweb}
\alias{inweb}
\title{Inweb title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 625641 rows and 3 columns.
}
\usage{
data("inweb")
}
\description{
Inweb title
}
\keyword{datasets}
