% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{mcode.fluff.complex}
\alias{mcode.fluff.complex}
\title{MCODE Fluff Complex}
\usage{
mcode.fluff.complex(graph, vertex.weight, fdt = 0.8, complex.g, seen)
}
\arguments{
\item{graph}{An `igraph` object representing the input graph.}

\item{vertex.weight}{A numeric vector containing weights for each vertex in the graph.}

\item{fdt}{A numeric threshold (default is 0.8) for neighborhood density. Neighbors with a density
exceeding this threshold will be incorporated into the complex.}

\item{complex.g}{A numeric vector containing the vertex indices that form the identified complex.}

\item{seen}{A numeric vector indicating the vertices that have been processed or visited.}
}
\value{
A numeric vector containing the vertex indices that constitute the expanded complex.
}
\description{
This function refines the identified complex resulting from the MCODE algorithm by considering
the neighborhood and density criteria. For each vertex in the identified complex, the function
examines its neighborhood. If the density of the neighborhood surpasses a specific threshold
(`fdt`), the neighboring vertices are incorporated into the complex.
}
\examples{
\dontrun{
# Assuming you have an igraph object (graph), vertex weights (vertex.weight),
# an initial complex (complex.g), and seen vertices (seen).
fluffed_complex <- mcode.fluff.complex(graph, vertex.weight, fdt=0.8, complex.g, seen)
}
}
