% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{mcode.vertex.weighting}
\alias{mcode.vertex.weighting}
\title{Calculate Vertex Weights Using MCODE Vertex Weighting}
\usage{
mcode.vertex.weighting(graph)
}
\arguments{
\item{graph}{An `igraph` object representing the input network. The function checks
whether the input is a valid `igraph` object.}
}
\value{
A numeric vector containing the weights of each vertex in the input graph.
        The order of weights corresponds to the order of vertices in the graph.
}
\description{
This function calculates the weights of vertices in a given graph based on the
k-core decomposition method used in MCODE (Molecular Complex Detection). The weights
reflect the relative importance or centrality of each vertex within the network,
aiding in community detection and other graph analysis tasks.
}
\examples{
\dontrun{
library(igraph)

# Creating a sample graph
g <- make_ring(10)

# Calculating vertex weights
weights <- mcode.vertex.weighting(g)

print(weights)
}
}
