% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nbrintersect}
\alias{nbrintersect}
\title{Length of intersect}
\usage{
nbrintersect(v1, v2)
}
\arguments{
\item{v1}{A vector}

\item{v2}{Another vector}
}
\value{
The length of the intersect between v1 and v2
}
\description{
Identical to length(intersect(x,y))
}
\examples{
ll = nbrintersect(1:10, 1:5)
}
