% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nbrsetdiff}
\alias{nbrsetdiff}
\title{Length of setdiff}
\usage{
nbrsetdiff(v1, v2)
}
\arguments{
\item{v1}{A vector}

\item{v2}{Another vector}
}
\value{
The length of the setdiff between v1 and v2
}
\description{
Identical to length(setdiff(x,y))
}
\examples{
ll = nbrsetdiff(1:10, 1:5)
}
