% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nbrunique}
\alias{nbrunique}
\title{Length of unique}
\usage{
nbrunique(v1)
}
\arguments{
\item{v1}{A vector}
}
\value{
The length of the unique between v1 and v2
}
\description{
Identical to length(unique(v1))
}
\examples{
ll = nbrunique(c(1,2,1))
}
