% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nbrunlist}
\alias{nbrunlist}
\title{Length of unlist}
\usage{
nbrunlist(v1)
}
\arguments{
\item{v1}{A list}
}
\value{
The length of the unlist between v1 and v2
}
\description{
Identical to length(unlist(v1))
}
\examples{
ll = nbrunlist(list(1:3, 2:4))
}
