% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_analysis.R
\name{ora_analysis}
\alias{ora_analysis}
\title{Over-Representation Analysis (ORA) on Network Communities}
\usage{
ora_analysis(
  graph,
  communities,
  organism,
  collection = NULL,
  ontology = NULL,
  custom_gene_set = F,
  universe,
  id_type
)
}
\arguments{
\item{graph}{An igraph object representing the network.}

\item{communities}{A list of igraph subgraphs representing the network communities.}

\item{organism}{A string indicating the organism, e.g., "human" or "mouse".}

\item{collection}{A string indicating the MSigDB collection, or NULL if not using MSigDB.}

\item{ontology}{A string indicating the MSigDB ontology, or NULL if not using this ontology.}

\item{custom_gene_set}{A logical flag, TRUE if using a custom gene set, FALSE otherwise. Defaults to FALSE.}

\item{universe}{A string indicating the universe for the ORA. Must be defined by the user. Must be entrezIDs (for now).}

\item{id_type}{A string indicating the type of gene ID used, e.g., "uniprot", "ensembl", etc.}
}
\value{
A list containing the ORA results for each community.
}
\description{
Performs over-representation analysis (ORA) on communities derived from a network graph.
It supports various types of gene IDs and can fetch gene sets from the Molecular Signatures Database (MSigDB)
or use custom gene sets.
}
\examples{
\dontrun{
# Using example igraph object for graph and dummy communities.
# This is a placeholder; replace with real examples.

# Assuming `sample_graph` is a predefined igraph object and `sample_communities` is a list of igraph subgraphs.
results <- ora_analysis(graph = sample_graph, communities = sample_communities, organism = "human", id_type = "uniprot")
}
}
