% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_heatmap.R
\name{ora_heatmap}
\alias{ora_heatmap}
\title{Create a heatmap for significant pathways}
\usage{
ora_heatmap(ora_results, significance_threshold = 0.05)
}
\arguments{
\item{ora_results}{A list of ORA results, with each list element representing the results for a specific community.}

\item{significance_threshold}{A numeric threshold value for significance. Defaults to 0.05.}
}
\value{
A heatmap visualizing the significant pathways across communities.
}
\description{
Generates a heatmap based on the Over-Representation Analysis (ORA) results to visualize
significant pathways across communities. Each row in the heatmap represents a unique significant pathway,
while columns represent different communities. A value of 1 denotes significance, whereas 0 denotes non-significance.
}
\examples{
\dontrun{
# Assuming `sample_ora_results` is a predefined list of ORA results.
ora_heatmap(ora_results = sample_ora_results, significance_threshold = 0.05)
}
}
