% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_filtering.R
\name{relevance_filtering}
\alias{relevance_filtering}
\title{Filter Nodes in an igraph Network Based on Relevance Score}
\usage{
relevance_filtering(network, cutoff = 0)
}
\arguments{
\item{network}{An igraph network object which must contain a vertex attribute named 'relevance_score'.}

\item{cutoff}{A numeric value specifying the threshold for the relevance score (default is 0).
Nodes with a relevance score below this cutoff will be removed.}
}
\value{
An igraph network object with nodes filtered based on the relevance score cutoff.
}
\description{
This function filters nodes in a given igraph network object based on the relevance score node attribute.
Nodes with relevance scores below the specified cutoff are removed from the network.
}
\examples{
\dontrun{
# Assuming 'net' is an igraph network object with a 'relevance_score' attribute for nodes:
filtered_net <- relevance_filtering(net, cutoff = 0.05)
}

}
