% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS_and_ML.R
\name{run_lasso}
\alias{run_lasso}
\title{Internal function: Do Adaptive LASSO for feature selection}
\usage{
run_lasso(index, Dataset, Label, nfolds, i = NULL, verbose = TRUE)
}
\arguments{
\item{index}{A numerical vector indicating which rows of the dataframe to work on}

\item{Dataset}{A numerical dataframe with features as columns and observations as rows}

\item{Label}{A boolean vector representing the class (TRUE/FALSE)}

\item{nfolds}{The number of folds during cross validation for tuning lambda}

\item{i}{Optional. Used during bootstraping}

\item{verbose}{Default to TRUE.Printing internal calculations.}
}
\value{
A string vector returning the selected features.
}
\description{
Return the list of features chosen through Adaptive LASSO during cross validation
}
