% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{RWR}
\alias{RWR}
\title{Internal function: random Walk function}
\usage{
RWR(
  Graph,
  Seeds,
  r = 0.5,
  attr = NULL,
  nCores = 1,
  adjacency = NULL,
  verbose = FALSE,
  stop_delta = 1e-10
)
}
\arguments{
\item{Graph}{An igraph object}

\item{Seeds}{Starting_nodes}

\item{r}{Restart probability}

\item{attr}{Edge weight, not functional yet.}

\item{nCores}{Default: 1 The number of cores used. Keep 1 if you want sequential calculations, >1 for parallel.}

\item{adjacency}{Optional. A precomputed adjacency matrix}

\item{verbose}{Wether to print internal calculations}

\item{stop_delta}{Optional. When to stop the random walks}
}
\value{
The length of the unlist between v1 and v2
}
\description{
Calculate a random walk on the network
}
