% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string}
\alias{string}
\title{String title}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13715404 rows and 3 columns.
}
\usage{
data("string")
}
\description{
String title
}
\keyword{datasets}
