% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{topo_metric_of_neighboring_signature}
\alias{topo_metric_of_neighboring_signature}
\title{Extract signature-based features}
\usage{
topo_metric_of_neighboring_signature(
  Graph,
  start_nodes,
  degs,
  fc_list,
  topo_results,
  order = 1,
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{degs}{A node signature (not list)}

\item{topo_results}{The results returned by the function extract_topolo}

\item{order}{The order of neighbors to consider}
}
\value{
Return something
}
\description{
Extract signature-based features from a biological network based on a list of signatures nodes.
}
