% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{topo_metric_with_foldchange}
\alias{topo_metric_with_foldchange}
\title{Extract signature-based features}
\usage{
topo_metric_with_foldchange(
  Graph,
  start_nodes,
  topo_results,
  fc_list,
  verbose = TRUE
)
}
\arguments{
\item{Graph}{A igraph object.}

\item{start_nodes}{Features are extracted for each of these nodes}

\item{topo_results}{The results returned by the function extract_topolo}
}
\value{
Return something
}
\description{
Extract signature-based features from a biological network based on a list of signatures nodes.
}
