% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOftranslate_database.R,
%   R/translate_database.R
\name{translate_database}
\alias{translate_database}
\title{Translate Database IDs}
\usage{
translate_database(database, id_type)

translate_database(database, id_type)
}
\arguments{
\item{database}{Either a character string specifying the type of database to translate - Accepted values are "inweb" or "string" - or a user-made
database with three columns, named "node", "interactor", "confidence_score".}

\item{id_type}{In case database is nor "string" or "inweb". A character stating which IDs are used in the submitted database.
Must be one among 'ensembl_peptide_id', 'ensembl_gene_id', 'ensembl_transcript_id', 'hgnc_symbol', 'uniprotswissprot'.}
}
\value{
A tibble with processed and translated databases enriched with multiple identifiers.
        The tibble contains columns with original identifiers as well as translated identifiers such as
        Ensembl Peptide IDs, UniProt IDs, Ensembl Gene IDs, Ensembl Transcript IDs, and HGNC symbols.

A tibble with processed and translated databases enriched with multiple identifiers.
        The tibble contains columns with original identifiers as well as translated identifiers such as
        Ensembl Peptide IDs, UniProt IDs, Ensembl Gene IDs, Ensembl Transcript IDs, and HGNC symbols.
}
\description{
This function translates and standardizes the given protein-protein interaction database ("inweb" or "string")
to ensure consistent usage of ID types across multiple databases. The databases are processed and enriched
with multiple identifiers like Ensembl Peptide ID, UniProt ID, Ensembl Gene ID, Ensembl Transcript ID, and HGNC symbols.

This function translates and standardizes the given protein-protein interaction database ("inweb" or "string")
to ensure consistent usage of ID types across multiple databases. The databases are processed and enriched
with multiple identifiers like Ensembl Peptide ID, UniProt ID, Ensembl Gene ID, Ensembl Transcript ID, and HGNC symbols.
}
\details{
The function performs several major tasks:
- Sets a timeout configuration for the fetching process.
- Validates the provided database type.
- Processes the chosen database by translating and standardizing identifiers.
- Handles missing and duplicated identifiers.
- Joins converted IDs to the original database.

The function might encounter issues such as missing IDs or duplicated mappings.
IDs that couldn’t be mapped to the database and IDs that returned multiple mappings
will be reported to help in troubleshooting and data cleaning.

The function performs several major tasks:
- Sets a timeout configuration for the fetching process.
- Validates the provided database type.
- Processes the chosen database by translating and standardizing identifiers.
- Handles missing and duplicated identifiers.
- Joins converted IDs to the original database.

The function might encounter issues such as missing IDs or duplicated mappings.
IDs that couldn’t be mapped to the database and IDs that returned multiple mappings
will be reported to help in troubleshooting and data cleaning.
}
\examples{
\dontrun{
# Example with the 'string' database
translated_string_db <- translate_database("string")

# Example with the 'inweb' database
translated_inweb_db <- translate_database("inweb")
}
\dontrun{
# Example with the 'string' database
translated_string_db <- translate_database("string")

# Example with the 'inweb' database
translated_inweb_db <- translate_database("inweb")
}
}
