% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_pulldown.R
\name{virtual_pulldown}
\alias{virtual_pulldown}
\title{Conduct Virtual Pull-Down Analysis on a Protein Network}
\usage{
virtual_pulldown(
  seed_nodes,
  database,
  id_type,
  string_confidence_score = 700,
  zs_confidence_score = 0.156,
  order = 1
)
}
\arguments{
\item{seed_nodes}{A vector of seed proteins.}

\item{database}{A preprocessed database for analysis, either derived from the
`translate_database` function or formatted similarly.}

\item{id_type}{The desired ID type for proteins. Can be one of: 'uniprot', 'ensg', 'ensp', 'enst', or 'hgnc'.}

\item{string_confidence_score}{A confidence score threshold for STRING database (default is 700). Should be set to NULL if not used.}

\item{zs_confidence_score}{ZS confidence score to use in the inweb database. Interactions with a score higher than 0.156 (default) are considered high confidence. Should be set to NULL if not used. Goes from 0 to 1.}

\item{order}{Numeric value (0 or 1) defining whether to fetch immediate interactors (order=1) or to also include second-level interactors (order=0).}
}
\value{
A list of two tibbles. One is the network and one is a node table, with two columns: seed and the calculated ecc_rel_score. The column "seed" describes
whether a node is a seed (1 for seed, 0 otherwise). Rows with NA values are removed from the final output,
and a message is displayed indicating the nodes that were removed.
}
\description{
This function conducts a virtual pull-down analysis on a protein-protein interaction network.
It fetches interactions from the given database (either 'inweb' or 'string'), converts them to the desired ID type,
and then identifies all possible interactions for the provided seed nodes. Additionally, it calculates relevance scores
for the identified interactions.
}
\examples{
\dontrun{
# Assuming you have a vector of seed nodes:
seeds <- c("P12345", "Q67890", ...)
result <- virtual_pulldown(seed_nodes = seeds, database = "string", id_type = "uniprot")
}
}
