% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vnames}
\alias{vnames}
\title{Retrieve node names}
\usage{
vnames(igraph_obj, pattern = NULL, negate = FALSE)
}
\arguments{
\item{igraph_obj}{A igraph object}

\item{pattern}{Pattern to look for.}

\item{negate}{If TRUE, return non-matching elements.}
}
\value{
The length of the unlist between v1 and v2
}
\description{
Retrieve the names of nodes from an igraph object
}
\examples{
# Don't run
# all_nodes = vnames(igraph_obj)
# protein_nodes = vnames(igraph_obj, type = "protein_")
}
